[FIX] Duplicated symbol link error on linking against static openal-soft.
The current openal.c works when linking against a shared libopenal.so.
However, when linking against a static libopenal.a, it causes multiple
`duplicated symbol' error.

The root cause is the way openal.c uses AL_IMPORT and ALC_IMPORT. The
macros defined a function pointer, with same name as the corresponding
extension functions in openal-soft. For example:

  #define AL_IMPORT(fun, t) t fun
  #define AL_IMPORT(fun, t) fun = (t)AlGetProcAddress(#fun)

So

  AL_IMPORT(alBufferDataStatic, PFNALBUFFERDATASTATICPROC)

is expanded to:

  PFNALBUFFERDATASTATICPROC alBufferDataStatic;
  alBufferDataStatic = (PFNALBUFFERDATASTATICPROC)alGetProcAddress("alBufferDataStatic")

So when linking against libopenal.a, there are two symbols with same
names, one is global function pointer, the other is function. They
conflicts and triggers link error.

I propose a fix that we add a suffix to global name, something like
below:

  #define AL_IMPORT(fun, t) t fun##_dyn
  #define AL_IMPORT(fun, t) fun##_dyn = (t)AlGetProcAddress(#fun)
  #define CHECK_EXT(fun) if(fun##_dyn == NULL) hl_error("Unsupported extension function")

which is expanded to

  PFNALBUFFERDATASTATICPROC alBufferDataStatic_dyn;
  alBufferDataStatic_dyn = (PFNALBUFFERDATASTATICPROC)alGetProcAddress("alBufferDataStatic")

We also need to add a macro to avoid calling *_dyn function pointers
directly:

  #define CALL(fun) fun##_dyn

So a function like below can reference to _dyn correctly.

  HL_PRIM void HL_NAME(al_buffer_data_static)(unsigned buffer, int format, vbyte *data, int len, int freq) {
	CHECK_EXT(alBufferDataStatic);
	CALL(alBufferDataStatic)(buffer, format, data, len, freq);
  }